# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bywise-node",
    "version": "0.0.1",
    "description": "bywise node",
    "contact": {
      "name": "Devel Blockchain",
      "email": "contact@develblockchain.com"
    }
  },
  "paths": {
...
    "/api/v2/blocks": {
      "get": {
        "security": [],
        "description": "Get blocks list",
        "tags": [
          "BlocksController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "where",
            "in": "query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockWhereQuery"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BlockWithStatusDTO"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      },
      "post": {
        "security": [],
        "description": "Insert new block",
        "tags": [
          "BlocksController"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
    "/api/v2/blocks/{hash}": {
      "get": {
        "security": [],
        "description": "Get block by hash",
        "tags": [
          "BlocksController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "pattern": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockWithStatusDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
...
    "/api/v2/slices": {
      "get": {
        "security": [],
        "description": "Get slices list",
        "tags": [
          "SlicesController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "where",
            "in": "query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceWhereQuery"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SliceDTO"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      },
      "post": {
        "security": [],
        "description": "Insert new slice",
        "tags": [
          "SlicesController"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SliceDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
    "/api/v2/slices/{hash}": {
      "get": {
        "security": [],
        "description": "Get slice by hash",
        "tags": [
          "SlicesController"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "pattern": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceDTO"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request"
          },
          "401": {
            "description": "Error: Unauthorized"
          },
          "422": {
            "description": "Error: Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockWhereQuery": {
        "example": "{\n  \"offset\": 0,\n  \"limit\": 100,\n  \"order\": \"desc\",\n  \"status\": \"mined\",\n  \"block.height\": 0,\n  \"block.slices\": {\n    \"$in\": [\n      \"0000000000000000000000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"block.chain\": \"mainnet\",\n  \"block.from\": \"BWS000000000000000000000000000000000000000000000\",\n  \"block.lastHash\": \"0000000000000000000000000000000000000000000000000000000000000000\",\n  \"block.hash\": \"0000000000000000000000000000000000000000000000000000000000000000\"\n}",
        "type": "object"
      },
      "SliceWhereQuery": {
        "example": "{\n  \"offset\": 0,\n  \"limit\": 100,\n  \"order\": \"desc\",\n  \"status\": \"mined\",\n  \"slice.height\": 0,\n  \"slice.chain\": \"mainnet\",\n  \"slice.transactions\": {\n    \"$in\": [\n      \"0000000000000000000000000000000000000000000000000000000000000000\"\n    ]\n  },\n  \"slice.from\": \"BWS000000000000000000000000000000000000000000000\",\n  \"slice.lastBlockHash\": \"0000000000000000000000000000000000000000000000000000000000000000\",\n  \"slice.hash\": \"0000000000000000000000000000000000000000000000000000000000000000\"\n}",
        "type": "object"
      },
...
      "SliceDTO": {
        "properties": {
          "height": {
            "type": "number"
          },
          "transactions": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "lastBlockHash": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sign": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BlockWithStatusDTO": {
        "properties": {
          "height": {
            "type": "number"
          },
          "slices": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "lastHash": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sign": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "externalTxID": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "BlockDTO": {
        "properties": {
          "height": {
            "type": "number"
          },
          "slices": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "version": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "lastHash": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sign": {
            "type": "string"
          },
          "externalTxID": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "example": "OK",
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "jwt": []
    }
  ]
}